+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version: 6.3
+// swift-tools-version: 6.4
 import PackageDescription
 
 let upcomingFeatures: [SwiftSetting] = [
@@ -15,11 +15,11 @@ let package = Package(
   name: "SwiftACD",
   defaultLocalization: "en",
   platforms: [
-    .macOS(.v15),
-    .iOS(.v18),
-    .watchOS(.v11),
-    .tvOS(.v18),
-    .visionOS(.v2)
+    .macOS(.v27),
+    .iOS(.v27),
+    .watchOS(.v27),
+    .tvOS(.v27),
+    .visionOS(.v27)
   ],
   products: [
     .library(name: "SwiftACD", targets: ["SwiftACD"]),